ref(preprod): Move Preprod{Static,Delta}GroupType to /preprod
